hugging-face-space/requirements.txt
openai
requests
beautifulsoup4
PyMuPDF
gradio
Pillow
reportlab
python-dotenv